+++ acir_field/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## [0.9.0](https://github.com/noir-lang/acvm/compare/acir_field-v0.8.1...acir_field-v0.9.0) (2023-04-07)
+
+
+### Miscellaneous Chores
+
+* **acir_field:** Synchronize undefined versions
+
+## [0.8.1](https://github.com/noir-lang/acvm/compare/acir_field-v0.8.0...acir_field-v0.8.1) (2023-03-30)
+
+
+### Miscellaneous Chores
+
+* **acir_field:** Synchronize undefined versions
+
+## [0.8.0](https://github.com/noir-lang/acvm/compare/acir_field-v0.7.1...acir_field-v0.8.0) (2023-03-28)
+
+
+### Miscellaneous Chores
+
+* **acir_field:** Synchronize undefined versions
+
+## [0.7.1](https://github.com/noir-lang/acvm/compare/acir_field-v0.7.0...acir_field-v0.7.1) (2023-03-27)
+
+
+### Miscellaneous Chores
+
+* **acir_field:** Synchronize undefined versions
+
+## [0.7.0](https://github.com/noir-lang/acvm/compare/acir_field-v0.6.0...acir_field-v0.7.0) (2023-03-23)
+
+
+### Miscellaneous Chores
+
+* **acir_field:** Synchronize undefined versions
+
 ## [0.6.0](https://github.com/noir-lang/acvm/compare/acir_field-v0.5.0...acir_field-v0.6.0) (2023-03-03)